Creates a data generator which generates labelled inputs for our model from the Lichess data.
This saves a lot of time as the RAM usage and the overall execution time is reduced significantly. It also gets rid of the need to create data before training, as data is created as it is needed. This also adds some improvements to the way the model is trained.